first cut at refactoring pdf2parquet
