+++ transforms/.make.cicd.targets
@@ -7,10 +7,15 @@
 include $(REPOROOT)/transforms/.make.transforms
 
 ######################################################################
-## Default setting for TRANSFORM_RUNTIME uses folder name-- Old layout
-TRANSFORM_PYTHON_RUNTIME_SRC_FILE=-m dpk_$(TRANSFORM_NAME).transform
-TRANSFORM_RAY_RUNTIME_SRC_FILE=-m dpk_$(TRANSFORM_NAME).ray.transform
-TRANSFORM_PYTHON_RUNTIME_SRC_FILE=-m dpk_$(TRANSFORM_NAME).spark.transform
+## Default setting for TRANSFORM_RUNTIME entry point: 
+# python -m dpk_html2parquet.ray.transform --help
+# or
+# python -m dpk_html2parquet.transform_python --help
+#
+TRANSFORM_PYTHON_SRC?="-m dpk_$(TRANSFORM_NAME).transform_python"
+TRANSFORM_RAY_SRC?="-m dpk_$(TRANSFORM_NAME).ray.transform"
+TRANSFORM_SPARK_SRC?="-m dpk_$(TRANSFORM_NAME).spark.transform"
+
 
 venv::	.defaults.create-venv
 	source venv/bin/activate && $(PIP) install -e $(REPOROOT)/data-processing-lib[ray,spark]
@@ -18,7 +23,6 @@ venv::	.defaults.create-venv
 	if [ -e requirements.txt ]; then \
 		source venv/bin/activate && $(PIP) install -r  requirements.txt; \
 	fi; 
-
 
 test:: .transforms.test-src test-image
 
@@ -28,62 +32,113 @@ clean:: .transforms.clean
 set-versions::
 
 ## We need to think how we want to do this going forward
-build::
 
-image::
-	@if [ -e Dockerfile ]; then \
-		$(MAKE) image-default ; \
-	else \
-		echo "Skipping image for $(shell pwd) since no Dockerfile is present"; \
+build:: image
+
+publish:
+	@if [ -e Dockerfile.python ]; then \
+		$(MAKE) DOCKER_REMOTE_IMAGE=$(DOCKER_REGISTRY_ENDPOINT)/$(TRANSFORM_NAME)-python:$(DOCKER_IMAGE_VERSION) \
+				.defaults.publish-image ; \
+	fi
+	@if [ -e Dockerfile.ray ]; then \
+		$(MAKE) DOCKER_REMOTE_IMAGE=$(DOCKER_REGISTRY_ENDPOINT)/$(TRANSFORM_NAME)-ray:$(DOCKER_IMAGE_VERSION) \
+				.defaults.publish-image ; \
+	fi
+	@if [ -e Dockerfile.spark ]; then \
+		$(MAKE) DOCKER_REMOTE_IMAGE=$(DOCKER_REGISTRY_ENDPOINT)/$(TRANSFORM_NAME)-spark:$(DOCKER_IMAGE_VERSION) \
+				.defaults.publish-image ; \
 	fi
 
-publish::
-	@if [ -e Dockerfile ]; then \
-		$(MAKE) publish-default ; \
-	else \
-		echo "Skipping publish for $(shell pwd) since no Dockerfile is present"; \
+test-image-sequence:: .defaults.lib-whl-image .transforms.test-image-help .transforms.clean
+
+test-image:: .default.build-lib-wheel
+	@if [ -e Dockerfile.python ]; then \
+		$(MAKE) DOCKER_FILE=Dockerfile.python \
+				TRANSFORM_RUNTIME_SRC_FILE=$(TRANSFORM_PYTHON_SRC) \
+				DOCKER_IMAGE_NAME=$(TRANSFORM_NAME)-python \
+				test-image-sequence ; \
+	fi
+	@if [ -e Dockerfile.ray ]; then \
+		$(MAKE) DOCKER_FILE=Dockerfile.ray \
+				TRANSFORM_RUNTIME_SRC_FILE=$(TRANSFORM_RAY_SRC) \
+				DOCKER_IMAGE_NAME=$(TRANSFORM_NAME)-ray \
+				BASE_IMAGE=$(RAY_BASE_IMAGE)  \
+				test-image-sequence ; \
+	fi
+	@if [ -e Dockerfile.spark ]; then \
+		$(MAKE) DOCKER_FILE=Dockerfile.spark \
+				TRANSFORM_RUNTIME_SRC_FILE=$(TRANSFORM_SPARK_SRC) \
+				DOCKER_IMAGE_NAME=$(TRANSFORM_NAME)-spark \
+				BASE_IMAGE=$(SPARK_BASE_IMAGE)  \
+				test-image-sequence ; \
+	fi
+	-rm -rf data-processing-dist
+
+
+image-python:
+	@if [ -e Dockerfile.python ]; then \
+		$(MAKE) DOCKER_FILE=Dockerfile.python \
+				DOCKER_IMAGE_NAME=$(TRANSFORM_NAME)-python \
+				.defaults.lib-whl-image ; \
 	fi
 
-publish-image::
-	@if [ -e Dockerfile ]; then \
-		$(MAKE) publish-image-default ; \
-	else \
-		echo "Skipping publish-image for $(shell pwd) since no Dockerfile is present"; \
+image-ray:
+	@if [ -e Dockerfile.ray ]; then \
+		$(MAKE) DOCKER_FILE=Dockerfile.ray \
+				DOCKER_IMAGE_NAME=$(TRANSFORM_NAME)-ray \
+				BASE_IMAGE=$(RAY_BASE_IMAGE)  \
+				.defaults.lib-whl-image ; \
 	fi
 
-test-image::
-	@if [ -e Dockerfile ]; then \
-		$(MAKE) test-image-default ; \
-	else \
-		echo "Skipping test-image for $(shell pwd) since no Dockerfile is present"; \
+image-spark:
+	@if [ -e Dockerfile.spark ]; then \
+		$(MAKE) DOCKER_FILE=Dockerfile.spark \
+				DOCKER_IMAGE_NAME=$(TRANSFORM_NAME)-spark \
+				BASE_IMAGE=$(SPARK_BASE_IMAGE)  \
+				.defaults.lib-whl-image ; \
 	fi
+
+image:: .default.build-lib-wheel
+	## Build all possible images unless a specific runtime is specified
+	@if [ -z "$(BUILD_SPECIFIC_RUNTIME)" ] || [ "$(BUILD_SPECIFIC_RUNTIME)" == "python" ]; then \
+		$(MAKE) image-python ; \
+	fi
+	@if [ -z "$(BUILD_SPECIFIC_RUNTIME)" ] || [ "$(BUILD_SPECIFIC_RUNTIME)" == "ray" ]; then \
+		$(MAKE) image-ray ; \
+	fi
+	@if [ -z "$(BUILD_SPECIFIC_RUNTIME)" ] || [ "$(BUILD_SPECIFIC_RUNTIME)" == "spark" ]; then \
+		$(MAKE) image-spark ; \
+	fi
+	-rm -rf data-processing-dist
 
 test-src:: .transforms.test-src
 
 setup:: .transforms.setup
 
-publish-default:: publish-image
+kind-load-image:: .transforms.kind-load-image
 
-publish-image-default:: .defaults.publish-image
-        
-test-image-default:: image .transforms.test-image-help .defaults.test-image-pytest .transforms.clean
+.PHONY: workflow-vent
+workflow-venv:
+	if [ -e kfp_ray ]; then                 \
+	    $(MAKE) -C kfp_ray TRANSFORM_NAME=$(TRANSFORM_NAME) workflow-venv;   \
+	fi
 
-build-lib-wheel:
-	make -C $(REPOROOT)/data-processing-lib build-pkg-dist
+.PHONY: workflow-test
+workflow-test:
+	if [ -e kfp_ray ]; then                 \
+	    $(MAKE) -C kfp_ray TRANSFORM_NAME=$(TRANSFORM_NAME) workflow-test;   \
+	fi
 
-image-default:: build-lib-wheel
-	@$(eval LIB_WHEEL_FILE := $(shell find $(REPOROOT)/data-processing-lib/dist/*.whl))
-	rm -fr dist && mv $(REPOROOT)/data-processing-lib/dist .
-	$(eval WHEEL_FILE_NAME := $(shell basename $(LIB_WHEEL_FILE)))
-	$(DOCKER) build -t $(DOCKER_IMAGE_NAME) $(DOCKER_BUILD_EXTRA_ARGS)	\
-		--platform $(DOCKER_PLATFORM)  			\
-		--build-arg EXTRA_INDEX_URL=$(EXTRA_INDEX_URL)	\
-		--build-arg BASE_IMAGE=$(RAY_BASE_IMAGE)  		\
-		--build-arg BUILD_DATE=$(shell date -u +'%Y-%m-%dT%H:%M:%SZ')  	\
-		--build-arg WHEEL_FILE_NAME=$(WHEEL_FILE_NAME) \
-		--build-arg TRANSFORM_NAME=$(TRANSFORM_NAME) \
-		--build-arg GIT_COMMIT=$(shell git log -1 --format=%h) .
-	$(DOCKER) tag $(DOCKER_LOCAL_IMAGE)  $(DOCKER_REMOTE_IMAGE)
-	rm -fr dist
+.PHONY: workflow-upload
+workflow-upload:
+	if [ -e kfp_ray ]; then                 \
+	    $(MAKE) -C kfp_ray TRANSFORM_NAME=$(TRANSFORM_NAME) workflow-upload; \
+	fi
+
+.PHONY: workflow-build
+workflow-build:
+	if [ -e kfp_ray ]; then                 \
+	    $(MAKE) -C  kfp_ray TRANSFORM_NAME=$(TRANSFORM_NAME) workflow-build; \
+	fi
