+++ transforms/language/pdf2parquet/dpk_pdf2parquet/transform_python.py
@@ -16,8 +16,8 @@ from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.runtime.pure_python.runtime_configuration import (
     PythonTransformRuntimeConfiguration,
 )
-from data_processing.utils import get_logger
-from pdf2parquet_transform import Pdf2ParquetTransformConfiguration
+from data_processing.utils import ParamsUtils, get_logger
+from dpk_pdf2parquet.transform import Pdf2ParquetTransformConfiguration
 
 
 logger = get_logger(__name__)
@@ -34,6 +34,31 @@ class Pdf2ParquetPythonTransformConfiguration(PythonTransformRuntimeConfiguratio
         :param base_configuration - base configuration class
         """
         super().__init__(transform_config=Pdf2ParquetTransformConfiguration())
+
+
+
+#Class used by the notebooks to ingest binary files and create parquet files
+class Pdf2ParquetRuntime():
+    def __init__(self, **kwargs):
+        self.params={}
+        for key in kwargs:
+            self.params[key]=kwargs[key]
+        # if input_folder and output_folder are specified, then assume it is represent data_local_config
+        try:
+            local_conf={k:self.params[k] for k in ('input_folder', 'output_folder')}
+            self.params['data_local_config']= ParamsUtils.convert_to_ast(local_conf)
+            del self.params['input_folder']
+            del self.params['output_folder']
+        except:
+            pass        
+    
+    def transform(self):
+        sys.argv = ParamsUtils.dict_to_req(d=(self.params))
+        # create launcher
+        launcher = PythonTransformLauncher(Pdf2ParquetPythonTransformConfiguration())
+        # launch
+        return_code = launcher.launch()
+        return return_code
 
 
 if __name__ == "__main__":
