--- transforms/language/pdf2parquet/ray/test-data/expected/metadata.json
@@ -1,64 +1,0 @@
-{
-  "pipeline": "pipeline_id",
-  "job details": {
-    "job category": "preprocessing",
-    "job name": "pdf2parquet",
-    "job type": "pure python",
-    "job id": "job_id",
-    "start_time": "2024-11-13 08:35:51",
-    "end_time": "2024-11-13 08:36:23",
-    "status": "success"
-  },
-  "code": {
-    "github": "github",
-    "commit_hash": "12345",
-    "path": "path"
-  },
-  "job_input_params": {
-    "batch_size": -1,
-    "artifacts_path": null,
-    "contents_type": "text/markdown",
-    "do_table_structure": true,
-    "do_ocr": true,
-    "ocr_engine": "easyocr",
-    "bitmap_area_threshold": 0.05,
-    "pdf_backend": "dlparse_v2",
-    "double_precision": 0,
-    "checkpointing": false,
-    "max_files": -1,
-    "random_samples": -1,
-    "files_to_use": [
-      ".pdf",
-      ".docx",
-      ".pptx",
-      ".zip"
-    ],
-    "num_processors": 0
-  },
-  "execution_stats": {
-    "cpus": 147.5,
-    "gpus": 0,
-    "memory": 33.72,
-    "object_store": 0,
-    "execution time, min": 0.522
-  },
-  "job_output_stats": {
-    "source_files": 2,
-    "source_size": 605137,
-    "result_files": 2,
-    "result_size": 33078,
-    "processing_time": 4.221,
-    "nrows": 3,
-    "nsuccess": 3,
-    "nfail": 0,
-    "nskip": 0
-  },
-  "source": {
-    "name": "/Users/dol/codes/data-prep-kit/transforms/language/pdf2parquet/python/test-data/input",
-    "type": "path"
-  },
-  "target": {
-    "name": "/Users/dol/codes/data-prep-kit/transforms/language/pdf2parquet/python/output",
-    "type": "path"
-  }
-}
